file-structure-optimising
