--- web/forms_team.py
@@ -1,1 +1,0 @@
-
