+++ web/templates/teams/create.html
@@ -1,97 +1,42 @@
 {% extends "base.html" %}
+{% load i18n %}
 
-{% block title %}Create Team Goal - Alpha One Labs{% endblock %}
+{% block title %}{% trans "Create Team Goal" %}{% endblock %}
 
 {% block content %}
-<main class="container mx-auto px-4 py-8">
+<div class="container mx-auto px-4 py-8">
     <div class="max-w-2xl mx-auto">
-        <div class="flex items-center justify-between mb-8">
-            <h1 class="text-3xl font-bold text-gray-900 dark:text-white">Create Team Goal</h1>
-            <a href="{% url 'team_goals' %}" 
-               class="text-blue-600 hover:text-blue-700 dark:text-blue-400 dark:hover:text-blue-300 flex items-center">
-                <i class="fas fa-arrow-left mr-2"></i>
-                Back to Goals
-            </a>
-        </div>
-
-        <div class="bg-white dark:bg-gray-800 rounded-lg shadow-lg p-6">
-            <form method="post" class="space-y-6">
-                {% csrf_token %}
-                
-                {% if form.non_field_errors %}
-                    <div class="bg-red-50 dark:bg-red-900/50 border border-red-200 dark:border-red-800 rounded-lg p-4 mb-6">
-                        {% for error in form.non_field_errors %}
-                            <p class="text-red-600 dark:text-red-400">{{ error }}</p>
-                        {% endfor %}
-                    </div>
+        <h1 class="text-2xl font-bold text-gray-800 dark:text-gray-200 mb-8">{% trans "Create Team Goal" %}</h1>
+        
+        <form method="post" class="bg-white dark:bg-gray-800 rounded-lg shadow-lg p-6">
+            {% csrf_token %}
+            
+            {% for field in form %}
+            <div class="mb-6">
+                <label for="{{ field.id_for_label }}" class="block text-sm font-medium text-gray-700 dark:text-gray-300 mb-2">
+                    {{ field.label }}
+                </label>
+                {{ field }}
+                {% if field.help_text %}
+                <p class="mt-1 text-sm text-gray-500 dark:text-gray-400">{{ field.help_text }}</p>
                 {% endif %}
-
-                <!-- Title Field -->
-                <div>
-                    <label for="{{ form.title.id_for_label }}" 
-                           class="block text-sm font-medium text-gray-700 dark:text-gray-300 mb-2">
-                        Title
-                    </label>
-                    {{ form.title }}
-                    {% if form.title.errors %}
-                        <div class="mt-1 text-sm text-red-600 dark:text-red-400">
-                            {{ form.title.errors|join:", " }}
-                        </div>
-                    {% endif %}
+                {% if field.errors %}
+                <div class="mt-1 text-red-600 dark:text-red-400 text-sm">
+                    {{ field.errors }}
                 </div>
-
-                <!-- Description Field -->
-                <div>
-                    <label for="{{ form.description.id_for_label }}" 
-                           class="block text-sm font-medium text-gray-700 dark:text-gray-300 mb-2">
-                        Description
-                    </label>
-                    {{ form.description }}
-                    {% if form.description.errors %}
-                        <div class="mt-1 text-sm text-red-600 dark:text-red-400">
-                            {{ form.description.errors|join:", " }}
-                        </div>
-                    {% endif %}
-                </div>
-
-                <!-- Target Date Field -->
-                <div>
-                    <label for="{{ form.target_date.id_for_label }}" 
-                           class="block text-sm font-medium text-gray-700 dark:text-gray-300 mb-2">
-                        Target Date
-                    </label>
-                    {{ form.target_date }}
-                    {% if form.target_date.errors %}
-                        <div class="mt-1 text-sm text-red-600 dark:text-red-400">
-                            {{ form.target_date.errors|join:", " }}
-                        </div>
-                    {% endif %}
-                </div>
-
-                <!-- Status Field -->
-                <div>
-                    <label for="{{ form.status.id_for_label }}" 
-                           class="block text-sm font-medium text-gray-700 dark:text-gray-300 mb-2">
-                        Status
-                    </label>
-                    {{ form.status }}
-                    {% if form.status.errors %}
-                        <div class="mt-1 text-sm text-red-600 dark:text-red-400">
-                            {{ form.status.errors|join:", " }}
-                        </div>
-                    {% endif %}
-                </div>
-
-                <!-- Submit Button -->
-                <div class="flex justify-end">
-                    <button type="submit" 
-                            class="inline-flex items-center px-6 py-3 bg-blue-600 hover:bg-blue-700 text-white rounded-lg transition duration-200">
-                        <i class="fas fa-save mr-2"></i>
-                        Create Team Goal
-                    </button>
-                </div>
-            </form>
-        </div>
+                {% endif %}
+            </div>
+            {% endfor %}
+            
+            <div class="flex justify-end mt-6">
+                <a href="{% url 'team_goals' %}" class="mr-4 px-6 py-2 bg-gray-100 hover:bg-gray-200 dark:bg-gray-700 text-gray-700 dark:text-gray-300 rounded-lg transition duration-200">
+                    {% trans "Cancel" %}
+                </a>
+                <button type="submit" class="bg-blue-600 hover:bg-blue-700 text-white px-6 py-2 rounded-lg transition duration-200">
+                    {% trans "Create Goal" %}
+                </button>
+            </div>
+        </form>
     </div>
-</main>
+</div>
 {% endblock %}
